+++ Src/KNote/Service/IKntUserService.cs
@@ -13,7 +13,7 @@ namespace KNote.Service
         Task<Result<UserDto>> GetByUserNameAsync(string userName);
         Task<Result<UserDto>> SaveAsync(UserDto entityInfo);
         Task<Result<UserDto>> DeleteAsync(Guid userId);
-        Task<Result<int>> GetCount();
+        Task<Result<long>> GetCount();
         Task<Result<UserDto>> Authenticate(string username, string password);        
         Task<Result<UserDto>> Create(UserRegisterDto userRegisterInfoDto);
     }
